feat: added test for localStorage, changed eslint config
