+++ .eslintrc.json
@@ -1,7 +1,9 @@
 {
   "env": {
     "browser": true,
-    "es2021": true
+    "es2021": true,
+    "node": true,
+    "jest": true
   },
   "extends": [
     "airbnb",
@@ -24,15 +26,17 @@
     "import/prefer-default-export": "off",
     "simple-import-sort/imports": "error",
     "simple-import-sort/exports": "error",
-    "unicorn/filename-case": [
+    "unicorn/filename-case": "off",
+    "unicorn/prevent-abbreviations": [
       "error",
       {
-        "cases": {
-          "camelCase": true,
-          "pascalCase": true
+        "replacements": {
+          "env": false
         }
       }
-    ]
+    ],
+    "func-names": "off",
+    "react/jsx-props-no-spreading": "off"
   },
   "settings": {
     "import/resolver": {
