+++ package.json
@@ -18,8 +18,7 @@
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.9.0",
-    "rimraf": "^4.4.0"
+    "react-router-dom": "^6.9.0"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
