+++ src/hoc/withRouter.tsx
@@ -1,16 +1,15 @@
 import React from 'react'
 import { useLocation, useNavigate, useParams } from 'react-router-dom'
 
-export interface WithRouterProperties {
+export type WithRouterProperties = {
   location: ReturnType<typeof useLocation>
-  params: Record<string, string>
+  params: ReturnType<typeof useParams>
   navigate: ReturnType<typeof useNavigate>
 }
 
 export const withRouter = <Properties extends WithRouterProperties>(
   Component: React.ComponentType<Properties>,
 ) => {
-  // eslint-disable-next-line func-names
   return function (properties: Omit<Properties, keyof WithRouterProperties>) {
     const location = useLocation()
     const parameters = useParams()
@@ -18,7 +17,6 @@ export const withRouter = <Properties extends WithRouterProperties>(
 
     return (
       <Component
-        // eslint-disable-next-line react/jsx-props-no-spreading
         {...(properties as Properties)}
         location={location}
         params={parameters}
